Add test to person_view_spec for adding a new email addres to the page when
the Add New Email button is clicked, and expect to reroute to the person page
afterward. Passed person_id from the url params to the form page, then to the new
action, and rails handled the handoff to the create action. Change controller tests
and let clauses to allow for the person_id to be there.